+++ AustinHarris.JsonRpcTestN/Test.cs
@@ -95,9 +95,9 @@ namespace AustinHarris.JsonRpcTestN
                 Tuple.Create ("sooper", typeof(string)),
                 Tuple.Create ("returns", typeof(string))
             }.ToDictionary(x => x.Item1, x => x.Item2);
-#pragma warning disable CS0618
+#pragma warning disable CS0618, JSONRPC0002
             h.RegisterFuction("workie", metadata, new System.Collections.Generic.Dictionary<string, object>(),new Func<string, string>(x => "workie ... " + x));
-#pragma warning restore CS0618
+#pragma warning restore CS0618, JSONRPC0002
 
             string request = @"{""method"":""workie"",""params"":{""sooper"":""good""},""id"":1}";
             string expectedResult = "{\"jsonrpc\":\"2.0\",\"result\":\"workie ... good\",\"id\":1}";
@@ -1614,9 +1614,9 @@ namespace AustinHarris.JsonRpcTestN
                 Tuple.Create ("sooper", typeof(string)),
                 Tuple.Create ("returns", typeof(string))
             }.ToDictionary(x => x.Item1, x => x.Item2);
-#pragma warning disable CS0618
+#pragma warning disable CS0618, JSONRPC0002
             h.RegisterFuction("workie", metadata, new System.Collections.Generic.Dictionary<string, object>(),new Func<string, string>(x => "workie ... " + x));
-#pragma warning restore CS0618
+#pragma warning restore CS0618, JSONRPC0002
 
             string request = @"{""method"":""workie"",""params"":{""sooper"":""good""},""id"":1}";
             string expectedResult = "{\"jsonrpc\":\"2.0\",\"result\":\"workie ... good\",\"id\":1}";
@@ -1856,9 +1856,9 @@ namespace AustinHarris.JsonRpcTestN
                 Tuple.Create ("sooper", typeof(string)),
                 Tuple.Create ("returns", typeof(string))
             }.ToDictionary(x => x.Item1, x => x.Item2);
-#pragma warning disable CS0618
+#pragma warning disable CS0618, JSONRPC0002
             h.RegisterFuction("workie", metadata, new System.Collections.Generic.Dictionary<string, object>(), new Func<string, string>(x => "workie ... " + x));
-#pragma warning restore CS0618
+#pragma warning restore CS0618, JSONRPC0002
 
             string request = @"{""method"":""workie"",""params"":{""sooper"":""good""},""id"":1}";
             string expectedResult = "{\"jsonrpc\":\"2.0\",\"result\":\"workie ... good\",\"id\":1}";